Move models-related commands to separate file
From `cli.py` to `groups/models.py`
